+++ src/Symfony/Component/DependencyInjection/EnvVarProcessorInterface.php
@@ -11,6 +11,8 @@
 
 namespace Symfony\Component\DependencyInjection;
 
+use Symfony\Component\DependencyInjection\Exception\RuntimeException;
+
 /**
  * The EnvVarProcessorInterface is implemented by objects that manage environment-like variables.
  *